requirements.txt
# Dependencias para Analizador de Patrones de Ruleta
# Python 3.12 recomendado (requerido para Super Resolution)

pyautogui
easyocr
pygame-ce
opencv-contrib-python==4.9.0.80
numpy<2
